move field `compatibleMovieTypes` to model HallType instead of Hall
