+++ models/HallType.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const Joi = require('joi');
+Joi.objectid = require('joi-objectid')(Joi);
 
 const hallTypeSchema = new mongoose.Schema({
     name: {
@@ -19,7 +20,13 @@ const hallTypeSchema = new mongoose.Schema({
     },
     updatedAt: {
         type: Date
-    }
+    },
+    compatibleMovieTypes: [
+        {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: 'MovieType'
+        }
+    ]
 });
 
 // Create `updatedAt` field
@@ -29,13 +36,15 @@ hallTypeSchema.pre('findOneAndUpdate', function () {
 
 const validationSchema = {
     name: Joi.string().max(50),
-    description: Joi.string()
+    description: Joi.string(),
+    compatibleMovieTypeIds: Joi.array().items(Joi.objectid()).min(1)
 };
 
 function validateOnCreateHallType(hallType) {
     const tmpValidationSchema = { ...validationSchema };
     tmpValidationSchema.name = tmpValidationSchema.name.required();
     tmpValidationSchema.description = tmpValidationSchema.description.required();
+    tmpValidationSchema.compatibleMovieTypeIds = tmpValidationSchema.compatibleMovieTypeIds.required();
     const schema = Joi.object(tmpValidationSchema);
 
     return schema.validate(hallType);
