+++ seeds/data/halls.json
@@ -5,11 +5,7 @@
         "seatRows": ["A", "B", "C", "D"],
         "seatColumns": [1, 2, 3, 4],
         "cinema": "5d713995b721c3bb38c1f5d0",
-        "hallType": "5f80169afe932e3d4055d1ea",
-        "compatibleMovieTypes": [
-            "5f84030ea795143ed451ddbf",
-            "5f8409065fc86e09e4752519"
-        ]
+        "hallType": "5f80169afe932e3d4055d1ea"
     },
     {
         "_id": "5f7e6fc7b2784c41dc7c7091",
@@ -17,11 +13,7 @@
         "seatRows": ["A", "B", "C", "D", "E", "F", "G", "H"],
         "seatColumns": [1, 2, 3, 4, 5, 6, 7],
         "cinema": "5d713995b721c3bb38c1f5d0",
-        "hallType": "5f80186e066d282d14640df0",
-        "compatibleMovieTypes": [
-            "5f84091350eab4485c327d1d",
-            "5f8409065fc86e09e4752519"
-        ]
+        "hallType": "5f80186e066d282d14640df0"
     },
     {
         "_id": "5f7e6fe396f1a141f0633a68",
@@ -29,7 +21,6 @@
         "seatRows": ["A", "B", "C", "D"],
         "seatColumns": ["I", "II", "III", "IV", "V"],
         "cinema": "5d713a66ec8f2b88b8f830b8",
-        "hallType": "5f80169afe932e3d4055d1ea",
-        "compatibleMovieTypes": ["5f84030ea795143ed451ddbf"]
+        "hallType": "5f80169afe932e3d4055d1ea"
     }
 ]
