+++ tests/integration/routes/hallTypes.test.js
@@ -1,6 +1,7 @@
 const request = require('supertest');
 const mongoose = require('mongoose');
 const { HallType } = require('../../../models/HallType');
+const { MovieType } = require('../../../models/MovieType');
 let server;
 
 describe('Hall Types', () => {
@@ -99,10 +100,34 @@ describe('Hall Types', () => {
     });
 
     describe('POST /api/v1/hall-types', () => {
+        let movieTypes;
+        let movieTypeIds;
+
         const data = {
             name: '4DX Hall',
             description: 'Equipped with motion and comfortable seats'
         };
+
+        beforeEach(async () => {
+            movieTypes = await MovieType.create([
+                {
+                    name: '2D',
+                    description: 'Simple 2D technology'
+                },
+                {
+                    name: '3D',
+                    description:
+                        'Exciting 3D technology with surrounding sounds'
+                }
+            ]);
+
+            movieTypeIds = movieTypes.map((m) => m._id);
+            data.compatibleMovieTypeIds = movieTypeIds;
+        });
+
+        afterEach(async () => {
+            await MovieType.deleteMany();
+        });
 
         const exec = () => request(server).post('/api/v1/hall-types');
 
@@ -163,6 +188,54 @@ describe('Hall Types', () => {
             expect(res.status).toBe(400);
         });
 
+        it('should return 400 if compatibleMovieTypeIds is not provided', async () => {
+            const curData = { ...data };
+            delete curData.compatibleMovieTypeIds;
+
+            const res = await exec().send(curData);
+
+            expect(res.status).toBe(400);
+        });
+
+        it('should return 400 if compatibleMovieTypeIds is an empty array', async () => {
+            const curData = { ...data };
+            curData.compatibleMovieTypeIds = [];
+
+            const res = await exec().send(curData);
+
+            expect(res.status).toBe(400);
+        });
+
+        it('should return 400 if compatibleMovieTypeIds is not an array of only string', async () => {
+            const curData = { ...data };
+            curData.compatibleMovieTypeIds = [1, 2];
+
+            const res = await exec().send(curData);
+
+            expect(res.status).toBe(400);
+        });
+
+        it('should return 400 if any id of compatibleMovieTypeIds is not a valid object Id', async () => {
+            const curData = { ...data };
+            curData.compatibleMovieTypeIds = [1, mongoose.Types.ObjectId()];
+
+            const res = await exec().send(curData);
+
+            expect(res.status).toBe(400);
+        });
+
+        it('should return 404 if any id of compatibleMovieTypeIds does not exist', async () => {
+            const curData = { ...data };
+            curData.compatibleMovieTypeIds = [
+                mongoose.Types.ObjectId(),
+                mongoose.Types.ObjectId()
+            ];
+
+            const res = await exec().send(curData);
+
+            expect(res.status).toBe(404);
+        });
+
         it('should return 201, and save the hall type if request is valid', async () => {
             const res = await exec().send(data);
 
@@ -182,6 +255,8 @@ describe('Hall Types', () => {
                 'description',
                 'Equipped with motion and comfortable seats'
             );
+            expect(res.body.data).toHaveProperty('compatibleMovieTypes');
+            expect(res.body.data.compatibleMovieTypes).toHaveLength(2);
         });
     });
 
@@ -201,7 +276,8 @@ describe('Hall Types', () => {
             await hallType.remove();
         });
 
-        const exec = () => request(server).put(`/api/v1/hall-types/${hallTypeId}`);
+        const exec = () =>
+            request(server).put(`/api/v1/hall-types/${hallTypeId}`);
 
         it('should return 404 if object ID is not valid', async () => {
             hallTypeId = 1;
@@ -215,57 +291,66 @@ describe('Hall Types', () => {
             const res = await exec().send({});
 
             expect(res.status).toBe(404);
-        })
+        });
 
         it('should return 400 if name is not a string', async () => {
             const res = await exec().send({ name: true });
 
             expect(res.status).toBe(400);
-        })
+        });
 
         it('should return 400 if name is an empty string', async () => {
             const res = await exec().send({ name: '' });
 
             expect(res.status).toBe(400);
-        })
+        });
 
         it('should return 400 if name is more than 100 characters', async () => {
             const name = new Array(102).join('a');
             const res = await exec().send({ name });
 
             expect(res.status).toBe(400);
-        })
+        });
 
         it('should return 400 if description is not a string', async () => {
             const res = await exec().send({ description: true });
 
             expect(res.status).toBe(400);
-        })
+        });
 
         it('should return 400 if description is an empty string', async () => {
             const res = await exec().send({ description: '' });
 
             expect(res.status).toBe(400);
-        })
+        });
 
         it('should return 200, and update the hall type if request is valid', async () => {
-            const res = await exec().send({ name: 'qwe', description: 'des-qwe' });
+            const res = await exec().send({
+                name: 'qwe',
+                description: 'des-qwe'
+            });
 
             const hallTypeInDb = await HallType.findById(hallTypeId);
 
             expect(res.status).toBe(200);
             expect(hallTypeInDb.name).toBe('qwe');
             expect(hallTypeInDb.description).toBe('des-qwe');
-        })
+        });
 
         it('should return 200, and return the updated hall type if request is valid', async () => {
-            const res = await exec().send({ name: 'qwe', description: 'des-qwe' });
+            const res = await exec().send({
+                name: 'qwe',
+                description: 'des-qwe'
+            });
 
             expect(res.status).toBe(200);
-            expect(res.body.data).toHaveProperty('_id', hallTypeId.toHexString());
+            expect(res.body.data).toHaveProperty(
+                '_id',
+                hallTypeId.toHexString()
+            );
             expect(res.body.data).toHaveProperty('name', 'qwe');
             expect(res.body.data).toHaveProperty('description', 'des-qwe');
-        })
+        });
     });
 
     describe('DELETE /api/v1/hall-types/:id', () => {
